Make TLS optional
This introduces a feature in Cargo.toml to enable TLS support to be
optional.  Various code has either been conditioned on the new feature,
or moved into the tls module if it's only ever used for TLS.
