+++ Cargo.toml
@@ -17,6 +17,10 @@ categories = ["api-bindings", "network-programming", "database-implementations"]
 
 license = "MIT/Apache-2.0"
 
+[features]
+default = ["tls"]
+tls = ["rustls"]
+
 [badges]
 azure-devops = { project = "jonhoo/jonhoo", pipeline = "msql-srv", build = "27" }
 codecov = { repository = "jonhoo/msql-srv", branch = "master", service = "github" }
@@ -28,7 +32,7 @@ mysql_common = "0.22"
 byteorder = "1"
 chrono = "0.4"
 time = "0.2.25"
-rustls = "0.20.0"
+rustls = {version = "0.20.0", optional=true}
 
 [dev-dependencies]
 postgres = "0.19.1"
